refactor(web): move roadmap hero utils to shared `_utils` folder
