+++ apps/web/app/(full-layout)/roadmaps/[id]/_utils/roadmap-hero-utils.ts
